+++ modules/atom/include/Chain.h
@@ -21,11 +21,15 @@ IMPATOM_BEGIN_NAMESPACE
 class IMPATOMEXPORT Chain : public Hierarchy {
 
   static void do_setup_particle(kernel::Model *m, kernel::ParticleIndex pi,
-                                char id) {
+                                std::string id) {
     m->add_attribute(get_id_key(), pi, id);
     if (!Hierarchy::get_is_setup(m, pi)) {
       Hierarchy::setup_particle(m, pi);
     }
+  }
+ static void do_setup_particle(kernel::Model *m, kernel::ParticleIndex pi,
+                                char c) {
+   do_setup_particle(m, pi, std::string(1, c));
   }
   static void do_setup_particle(kernel::Model *m, kernel::ParticleIndex pi,
                                 Chain o) {
@@ -34,6 +38,7 @@ class IMPATOMEXPORT Chain : public Hierarchy {
 
  public:
   IMP_DECORATOR_METHODS(Chain, Hierarchy);
+  IMP_DECORATOR_SETUP_1(Chain, std::string, id);
   IMP_DECORATOR_SETUP_1(Chain, char, id);
   IMP_DECORATOR_SETUP_1(Chain, Chain, other);
 
@@ -43,18 +48,33 @@ class IMPATOMEXPORT Chain : public Hierarchy {
   }
 
   //! Return the chain id
-  char get_id() const {
-    return static_cast<char>(
-        get_model()->get_attribute(get_id_key(), get_particle_index()));
+  std::string get_id() const {
+    return get_model()->get_attribute(get_id_key(), get_particle_index());
+  }
+
+  //! Return the chain id
+  /** \deprecated_at{2.2} Use the string version. */
+  IMPATOM_DEPRECATED_FUNCTION_DECL(2.2)
+  char get_id_char() const {
+    return get_model()->get_attribute(get_id_key(), get_particle_index())[0];
   }
 
   //! Set the chain id
+  /** \deprecated_at{2.2} Use the string version instead. */
+  IMPATOM_DEPRECATED_FUNCTION_DECL(2.1)
   void set_id(char c) {
+    IMPATOM_DEPRECATED_FUNCTION_DEF(2.1, "Use a string to identify the chain");
+    get_model()->set_attribute(get_id_key(), get_particle_index(),
+                               std::string(1, c));
+  }
+
+  //! Set the chain id
+  void set_id(std::string c) {
     get_model()->set_attribute(get_id_key(), get_particle_index(), c);
   }
 
   //! The key used to store the chain
-  static IntKey get_id_key();
+  static StringKey get_id_key();
 };
 
 IMP_DECORATORS(Chain, Chains, Hierarchies);