+++ modules/atom/include/hierarchy_tools.h
@@ -95,7 +95,13 @@ IMPATOMEXPORT Hierarchy
 IMPATOMEXPORT std::string get_molecule_name(Hierarchy h);
 IMPATOMEXPORT Ints get_residue_indexes(Hierarchy h);
 IMPATOMEXPORT ResidueType get_residue_type(Hierarchy h);
-IMPATOMEXPORT int get_chain_id(Hierarchy h);
+IMPATOMEXPORT std::string get_chain_id(Hierarchy h);
+/** \deprecated_at{2.2} Use the string version. */
+IMPATOM_DEPRECATED_FUNCTION_DECL(2.2)
+inline char get_chain_id_char(Hierarchy h) {
+  IMPATOM_DEPRECATED_FUNCTION_DEF(2.2, "Use string version");
+  return get_chain_id(h)[0];
+}
 IMPATOMEXPORT AtomType get_atom_type(Hierarchy h);
 IMPATOMEXPORT std::string get_domain_name(Hierarchy h);
 IMPATOMEXPORT int get_copy_index(Hierarchy h);