+++ modules/rmf/src/atom_links.cpp
@@ -34,7 +34,7 @@ std::string get_good_name(kernel::Model *m, kernel::ParticleIndex h) {
     oss << atom::Residue(m, h).get_index();
     return oss.str();
   } else if (atom::Chain::get_is_setup(m, h)) {
-    return std::string(1, atom::Chain(m, h).get_id());
+    return atom::Chain(m, h).get_id();
   } else {
     return m->get_particle_name(h);
   }